Feat: admin인증 추가
- 다른 서버에있는 admin인증 api를 이용한 admin 인증 추가

+++ src/main/java/com/udoolleh/backend/web/CourseController.java
@@ -1,9 +1,14 @@
 package com.udoolleh.backend.web;
 
+import com.udoolleh.backend.exception.errors.CustomJwtRuntimeException;
+import com.udoolleh.backend.provider.security.JwtAuthTokenProvider;
+import com.udoolleh.backend.provider.service.AdminAuthenticationService;
 import com.udoolleh.backend.provider.service.CourseService;
 import com.udoolleh.backend.web.dto.CommonResponse;
 import com.udoolleh.backend.web.dto.RequestCourse;
 import com.udoolleh.backend.web.dto.ResponseCourse;
+import java.util.Optional;
+import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,9 +21,16 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CourseController {
     private final CourseService courseService;
+    private final JwtAuthTokenProvider jwtAuthTokenProvider;
+    private final AdminAuthenticationService adminAuthenticationService;
 
     @PostMapping("/course")
-    public ResponseEntity<CommonResponse> registerCourse(@Valid @RequestBody RequestCourse.RegisterCourseDto requestDto){
+    public ResponseEntity<CommonResponse> registerCourse(HttpServletRequest request, @Valid @RequestBody RequestCourse.RegisterCourseDto requestDto){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
+
         courseService.registerCourse(requestDto);
 
         return ResponseEntity.ok().body(CommonResponse.builder()
@@ -36,7 +48,12 @@ public class CourseController {
                 .build());
     }
     @DeleteMapping("/course/{id}")
-    public ResponseEntity<CommonResponse> deleteCourse(@PathVariable Long id){
+    public ResponseEntity<CommonResponse> deleteCourse(HttpServletRequest request, @PathVariable Long id){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
+
         courseService.deleteCourse(id);
 
         return ResponseEntity.ok().body(CommonResponse.builder()
